fix(snippets): Move snippet choice classes to match upstream
